# Conflicts:
#	jjikeobang-client/src/components/header/Profile.js
#	jjikeobang-client/src/pages/Voting.js

+++ jjikeobang-server/src/main/java/com/jjikeobang/member/service/MemberService.java
@@ -9,7 +9,8 @@ public interface MemberService {
     List<Member> selectAllMember();
 
     void insertMember(JoinMemberDTO member);
-    Member findById(int memberId);
+  
+    Member findById(Long memberId);
 
     Member findByLoginId(String userId);
